Added summary.py
summary.py will calculate the best/average fitnesses to be printed once the evolution is done. Main will need to be updated to use this later (once I've tested this probably, honestly). Also includes (untested) visualization graphs for the average and best fitness, and will later include a visualization for Routes.